requirements.txt
pyusb
setuptools<82
python-escpos>=3.1
pillow>=10.0
requests>=2.31
rapidfuzz>=3.0
rich>=13.0
flask>=3.0
